Another refinement is to define a `Usage()` exception, which we catch in an except clause at the end of `main()`:
@show main.py

This gives the `main()` function a single exit point, which is preferable over
multiple `return 2` statements. This also makes it easier to refactor the
argument parsing: `raise Usage` works just fine from inside a helper function,
but `return 2` would require careful passing on of the return values.

You might think that taking this to the extreme would move the `try`/`except`
clause out of the `main()` function, into the code at the end of the module (if
`__name__ == "__main__": ....` But that would mean that when you call `main()`
interactively, you'd get a traceback for command line syntax errors, which
isn't very helpful.

However, another generalization can be helpful: define another exception,
perhaps called `Error`, which is treated just like `Usage` but `return`s `1`.
This can then be used for expected errors like failure to open necessary files,
which are not command line syntax errors, but yet expected, and where again a
traceback doesn't feel very friendly.