Test robustness: Ignore changes to whitespace in markdown output
Different versions of python-markdown produce different whitespace.  This
defeats my naive tests which just do a diff.  This commit makes diff
ignore whitespace changes to prevent the tests breaking for no reason in
the future.

Unfortunately, even with diff flags `-B` and `-u` diff still can't cope
with an addition of a newline in the middle of some text so I've had to
make a change to the expected output of the `test_diff` test.  A fix for
this would be to normalise the expected output and the markdown output for
instance with html tidy before doing the file compare or using wdiff for a
word-based diff independent of newlines but both approaches would require
additional dependencies and are probably not worth it.